Biginteger befejezés

+++ src/main/Main.java
@@ -5,12 +5,14 @@ import encryption.EncryptionService;
 import input.InputHandler;
 import keygen.KeyGenService;
 
+import java.math.BigInteger;
+
 public class Main {
 
     private static Tasks chooseTask() {
         System.out.println("Milyen műveletet szeretne végezni? (titkosítás: 1, visszafejtés: 2, kulcsgenerálás: 3)");
-        int taskNum = InputHandler.getTaskInput("Művelet");
-        return Tasks.values()[taskNum - 1];
+        BigInteger taskNum = InputHandler.getTaskInput("Művelet");
+        return Tasks.values()[taskNum.intValue() - 1];
     }
 
     public static void main(String[] args) {
